requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==1.11.8
django-filter==2.1.0
djangorestframework==3.9.4
get==2019.4.13
idna==2.8
Markdown==3.1.1
Pillow==6.0.0
pkg-resources==0.0.0
post==2019.4.13
public==2019.4.13
pytz==2019.1
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
urllib3==1.25.3
